Tidy up admin commands

+++ bsrocks/commands/admin/fetch.lua
@@ -1,5 +1,7 @@
 local download = require "bsrocks.downloaders"
 local fileWrapper = require "bsrocks.lib.files"
+local log = require "bsrocks.lib.utils".log
+local patchspec = require "bsrocks.rocks.patchspec"
 local rockspec = require "bsrocks.rocks.rockspec"
 local serialize = require "bsrocks.lib.serialize"
 local settings = require "bsrocks.lib.settings"
@@ -7,51 +9,71 @@ local settings = require "bsrocks.lib.settings"
 local patchDirectory = settings.patchDirectory
 local servers = settings.servers
 
-local function execute(name, version)
-	if not name then error("Expected name", 0) end
+local function execute(...)
+	local patched, force
+	if select("#", ...) == 0 then
+		force = false
+		patched = patchspec.getAll()
+	else
+		force = true
+		for _, name in pairs({...}) do
+			local file = fs.combine(patchDirectory, name .. ".patchspec")
+			if not fs.exists(file) then error("No such patchspec " .. name, 0) end
 
-	local server, manifest = rockspec.findRock(servers, name)
-	if not server then
-		error("Cannot find '" .. name .. "'", 0)
+			patched[name] = serialize.unserialize(fileWrapper.read(file))
+		end
 	end
 
-	if not version then
-		version = rockspec.latestVersion(manifest, name)
+	local changed = false
+	for name, patchspec in pairs(patched) do
+		local dir = fs.combine(patchDirectory, "rocks-original/" .. name)
+		if force or not fs.isDir(dir) then
+			changed = true
+			log("Fetching " .. name)
+
+			fs.delete(dir)
+
+			local version = patchspec.version
+			if not patchspec.version then
+				error("Patchspec" .. name .. " has no version", 0)
+			end
+
+			local server, manifest = rockspec.findRock(servers, name)
+			if not server then
+				error("Cannot find '" .. name .. "'", 0)
+			end
+
+			local rock = rockspec.fetchRockspec(server, name, patchspec.version)
+
+			local files = rockspec.extractFiles(rock)
+			if #files == 0 then error("No files for " .. name .. "-" .. version, 0) end
+
+			local downloaded = download(rock.source, files)
+
+			if not downloaded then error("Cannot find downloader for " .. rock.source.url, 0) end
+
+			for name, contents in pairs(downloaded) do
+				fileWrapper.write(fs.combine(dir, name), contents)
+			end
+
+			fs.delete(fs.combine(patchDirectory, "rocks-changes/" .. name))
+		end
 	end
 
-	local rock = rockspec.fetchRockspec(server, name, version)
-
-	local files = rockspec.extractFiles(rock)
-	if #files == 0 then error("No files for " .. name .. "-" .. version, 0) end
-
-	local downloaded = download(rock.source, files)
-
-	if not downloaded then error("Cannot find downloader for " .. rock.source.url, 0) end
-
-	local dir = fs.combine(patchDirectory, "rocks-original/" .. name)
-	for name, contents in pairs(downloaded) do
-		fileWrapper.write(fs.combine(dir, name), contents)
+	if not changed then
+		error("No packages to fetch", 0)
 	end
 
-	fs.delete(fs.combine(patchDirectory, "rocks-changes/" .. name))
-
-	local info = fs.combine(patchDirectory, "rocks/" .. name .. ".patchspec")
-	local data = {}
-	if fs.exists(info) then
-		data = serialize.unserialize(fileWrapper.read(info))
-	end
-
-	data.version = version
-	data.dependencies = rock.dependencies
-	if not data.type then data.type = "patch" end
-	fileWrapper.write(info, serialize.serialize(data))
-
-	print("Run 'apply-patches " .. name .. "' to apply")
+	print("Run 'apply-patches' to apply")
 end
 
+local description = [[
+  [name] The name of the package to fetch. Otherwise all un-fetched packages will be fetched.
+]]
 return {
 	name = "fetch",
 	help = "Fetch a package for patching",
-	syntax = "<name> [version]",
-	execute = execute
+	syntax = "[name]...",
+	description = description,
+	execute = execute,
 }
